split dummy explainer
