+++ torch_geometric/explainability/algo/base.py
@@ -9,13 +9,10 @@ from torch_geometric.explainability.explanations import Explanation
 class ExplainerAlgorithm(torch.nn.Module):
     r"""An abstract class for explainer algorithms."""
     def __init__(self) -> None:
-        # TODO: check how to initialize... I don't like the fact that
-        # an explainer algorithm "has" a model
-        # it should be decoupled no ?
         super().__init__()
 
         # TODO: check if this is good practice, took the idea from captum
-        objective: Callable[
+        self.objective: Callable[
             [Tuple[torch.Tensor,
                    ...], torch.Tensor, torch.Tensor], torch.Tensor]
         r"""
@@ -50,8 +47,7 @@ class ExplainerAlgorithm(torch.nn.Module):
         model: torch.nn.Module,
         **kwargs,
     ) -> Explanation:
-        r"""This method computes the explanation of the GNN and return
-        a list of masks.
+        r"""This method computes the explanation of the GNN.
 
         Args:
             inputs (Tuple[torch.Tensor, ...]): the inputs of the GNN.
@@ -60,7 +56,11 @@ class ExplainerAlgorithm(torch.nn.Module):
             name_inputs (List[str]): the name of the inputs of the GNN.
                 (used to create the explanation object)
         """
-        pass
+
+    def forward(self, *x: torch.Tensor, edge_index: torch.Tensor,
+                target: torch.Tensor, model: torch.nn.Module,
+                **kwargs) -> Explanation:
+        return self.explain(x, edge_index, target, model, **kwargs)
 
     def supports(
         self,
@@ -78,40 +78,3 @@ class ExplainerAlgorithm(torch.nn.Module):
 
         """
         return True
-
-
-class RandomExplainer(ExplainerAlgorithm):
-    """Dummy explainer."""
-    def __init__(self) -> None:
-        super().__init__()
-
-    def explain(
-        self,
-        x: torch.Tensor,
-        edge_index: torch.Tensor,
-        target: torch.Tensor,
-        model: torch.nn.Module,
-        **kwargs,
-    ) -> Explanation:
-
-        node_features_mask = torch.rand(x.shape)
-        node_mask = torch.rand(x.shape[0])
-        edge_mask = torch.rand(edge_index.shape[1])
-        if "edge_attr" in kwargs:
-            edge_features_mask = torch.rand(kwargs["edge_attr"].shape)
-        return Explanation(
-            x=x,
-            edge_index=edge_index,
-            node_features_mask=node_features_mask,
-            node_mask=node_mask,
-            edge_mask=edge_mask,
-            edge_features_mask=edge_features_mask,
-            **kwargs,
-        )
-
-    def supports(
-        self,
-        explanation_type: str,
-        mask_type: str,
-    ) -> bool:
-        return mask_type != "layers"
